removed print statements, moved forms to forms.py, renamed functions.py to utils.py, generated random secret key, moved sensitive data to environment, set up sensitive data in Heroku environment, deleted C9 git files
