+++ .gitignore
@@ -1,1 +1,3 @@
 .env
+controllers.py
+controllers.pyc
