+++ data/countries.json
@@ -5,7 +5,7 @@
 {"name": "Albania", "code": "AL"},
 {"name": "Algeria", "code": "DZ"},
 {"name": "American Samoa", "code": "AS"},
-{"name": "AndorrA", "code": "AD"},
+{"name": "Andorra", "code": "AD"},
 {"name": "Angola", "code": "AO"},
 {"name": "Anguilla", "code": "AI"},
 {"name": "Antarctica", "code": "AQ"},
